pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["src*"]

[project]
name = "openecho"
version = "0.1.0"
description = "AI partner Telegram bot with modular skill system"
requires-python = ">=3.11"
dependencies = [
    "aiogram>=3.0",
    "redis>=5.0",
    "structlog>=24.0",
    "pyyaml>=6.0",
    "anthropic>=0.30",
    "python-dotenv>=1.0",
    "fastapi>=0.100",
    "uvicorn>=0.30",
    "websockets>=12.0",
    "httpx>=0.27",
    "chromadb>=0.5",
    "deepgram-sdk>=3.0",
    "todoist-api-python>=2.0",
    "croniter>=2.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "pytest-mock>=3.0",
    "ruff>=0.5",
]

[tool.pytest.ini_options]
markers = [
    "unit: fast tests with mocks",
    "integration: tests requiring Redis/API/ChromaDB",
]
asyncio_mode = "auto"

[tool.ruff]
line-length = 100
target-version = "py311"
